requirements.txt
django-ninja==0.13.0
django-money==2.0